refactor(tui::ui::components::popups::quit): move function "mount_quit_popup" to popup file
